Ultimate algorithm.  Works, but dead end.
New algorithm for more efficient exploration.  Divide unexplored area into small regions of 50 to 100 tiles each, like a patchwork quilt.  Call each patch a "goal".  Food, enemy hills, and clusters of enemy ants are also goals.  Assign just enough ants to each goal to handle it.  Something like 1 for an unexplored area, 1 for a food, 2 to 4 per enemy ant in a cluster, 10 for a hill.

Problem: requires a BFS for each goal, could be a zillion goals.  Solution: just make it really fast and do the zillion BFSs.

Real problem: seriously, it's hard to predict the number of goals.  Hard to guess whether it will time out sporadically in a big map with lots of enemies around.  It stresses me out worrying about timeouts, it goes against my design principles for this project, it casts doubt on this development direction, and it saps my will to pursue it.

Other real problem: it's still buggy, and it's hard to create a single visualization for an arbitrary number of distance maps.  So debugging is a slog.

I'm abandoning this branch to go back to my last submission (plus smart foraging).  This fast multiple-BFS implementation will be useful for speeding up the foraging code.